+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Optional namespacing for aggregate/projection streams and events allowing for isolation
   between applications. [#12](https://github.com/bilus/akasha/pull/12)
+* Fix Unhandled events in stream break aggregate loading. [Issue #5](https://github.com/bilus/akasha/issues/5)
 
 
 ## Version 0.3.0